app/requirements.txt
cassandra-driver==3.29.2
click==8.1.8
cramjam==2.10.0
fastparquet==2024.2.0
fsspec==2025.3.0
geomet==0.2.1.post1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pathvalidate==3.2.1
py4j==0.10.9.7
pyarrow==17.0.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
venv-pack==0.2.0
